Move SpecimenManager
