+++ study/api-src/org/labkey/api/study/StudyUtils.java
@@ -1,5 +1,7 @@
 package org.labkey.api.study;
 
+import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.dialect.SqlDialect;
 import org.labkey.api.util.DateUtil;
 
 import java.util.Calendar;
@@ -13,5 +15,16 @@ public class StudyUtils
     {
         Calendar cal = DateUtil.newCalendar(d.getTime());
         return cal.get(Calendar.YEAR) * 10000 + (cal.get(Calendar.MONTH) + 1) * 100 + cal.get(Calendar.DAY_OF_MONTH);
+    }
+
+    public static SQLFragment sequenceNumFromDateSQL(String dateColumnName)
+    {
+        // Returns a SQL statement that produces a single number from a date, in the form of YYYYMMDD.
+        SqlDialect dialect = StudyService.get().getStudySchema().getSqlDialect();
+        SQLFragment sql = new SQLFragment();
+        sql.append("(10000 * ").append(dialect.getDatePart(Calendar.YEAR, dateColumnName)).append(") + ");
+        sql.append("(100 * ").append(dialect.getDatePart(Calendar.MONTH, dateColumnName)).append(") + ");
+        sql.append("(").append(dialect.getDatePart(Calendar.DAY_OF_MONTH, dateColumnName)).append(")");
+        return sql;
     }
 }
