+++ study/src/org/labkey/study/controllers/specimen/ParticipantCommentAction.java
@@ -24,6 +24,7 @@ import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.User;
 import org.labkey.api.security.permissions.InsertPermission;
 import org.labkey.api.security.permissions.ReadPermission;
+import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.model.SpecimenComment;
@@ -32,7 +33,6 @@ import org.labkey.api.view.DataView;
 import org.labkey.api.view.InsertView;
 import org.labkey.api.view.NavTree;
 import org.labkey.api.view.UpdateView;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.controllers.InsertUpdateAction;
 import org.labkey.study.model.StudyImpl;
 import org.springframework.validation.BindException;
@@ -70,10 +70,10 @@ public abstract class ParticipantCommentAction extends InsertUpdateAction<Specim
                 Vial vial = SpecimenManagerNew.get().getVial(container, user, rowId);
                 if (vial != null)
                 {
-                    SpecimenComment comment = SpecimenManager.getInstance().getSpecimenCommentForVial(vial);
+                    SpecimenComment comment = SpecimenManager.get().getSpecimenCommentForVial(vial);
                     if (comment != null)
                     {
-                        SpecimenManager.getInstance().setSpecimenComment(user, vial, null,
+                        SpecimenManager.get().setSpecimenComment(user, vial, null,
                                 comment.isQualityControlFlag(), comment.isQualityControlFlagForced());
                     }
                 }
