+++ study/src/org/labkey/study/specimen/report/participant/ParticipantSiteReportFactory.java
@@ -17,13 +17,13 @@ package org.labkey.study.specimen.report.participant;
 
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.query.FieldKey;
+import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.Visit;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.Pair;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.controllers.specimen.SpecimenController;
 import org.labkey.study.specimen.report.SpecimenVisitReport;
 import org.labkey.study.specimen.report.SpecimenVisitReportParameters;
@@ -44,12 +44,12 @@ public class ParticipantSiteReportFactory extends SpecimenVisitReportParameters
     @Override
     protected List<? extends SpecimenVisitReport> createReports()
     {
-        List<? extends Visit> visits = SpecimenManager.getInstance().getVisitsWithSpecimens(getContainer(), getUser(), getCohort());
+        List<? extends Visit> visits = SpecimenManager.get().getVisitsWithSpecimens(getContainer(), getUser(), getCohort());
         List<ParticipantVisitReport> reports = new ArrayList<>();
         Set<LocationImpl> enrollmentLocations;
         if (_enrollmentSiteId == null)
         {
-            enrollmentLocations = SpecimenManager.getInstance().getEnrollmentSitesWithSpecimens(getContainer(), getUser());
+            enrollmentLocations = SpecimenManager.get().getEnrollmentSitesWithSpecimens(getContainer(), getUser());
             // add null to the set so we can search for ptid without an enrollment site:
             enrollmentLocations.add(null);
         }
@@ -92,7 +92,7 @@ public class ParticipantSiteReportFactory extends SpecimenVisitReportParameters
     public List<Pair<String, HtmlString>> getAdditionalFormInputHtml()
     {
         List<Pair<String, HtmlString>> inputs = new ArrayList<>(super.getAdditionalFormInputHtml());
-        Set<LocationImpl> locations = SpecimenManager.getInstance().getEnrollmentSitesWithSpecimens(getContainer(), getUser());
+        Set<LocationImpl> locations = SpecimenManager.get().getEnrollmentSitesWithSpecimens(getContainer(), getUser());
         // add null to the set so we can search for ptid without an enrollment site:
         locations.add(null);
         inputs.add(getEnrollmentSitePicker("enrollmentSiteId", locations, _enrollmentSiteId));
