+++ study/src/org/labkey/study/view/specimen/requestSpecimens.jsp
@@ -16,6 +16,7 @@
  */
 %>
 <%@ page import="org.labkey.api.data.Container"%>
+<%@ page import="org.labkey.api.specimen.SpecimenManager"%>
 <%@ page import="org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput"%>
 <%@ page import="org.labkey.api.specimen.Vial"%>
 <%@ page import="org.labkey.api.specimen.location.LocationImpl"%>
@@ -23,10 +24,9 @@
 <%@ page import="org.labkey.api.specimen.settings.SettingsManager"%>
 <%@ page import="org.labkey.api.study.SpecimenService"%>
 <%@ page import="org.labkey.api.view.HttpView"%>
-<%@ page import="org.labkey.api.view.JspView"%>
+<%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.ViewContext" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
-<%@ page import="org.labkey.study.SpecimenManager" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.CreateSpecimenRequestForm" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.HandleCreateSpecimenRequestAction" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.NewRequestBean" %>
@@ -51,7 +51,7 @@
     Container c = getContainer();
     List<LocationImpl> locations = LocationManager.get().getValidRequestingLocations(c);
     boolean shoppingCart = SettingsManager.get().isSpecimenShoppingCartEnabled(c);
-    boolean hasExtendedRequestView = SpecimenManager.getInstance().getExtendedSpecimenRequestView(context) != null;
+    boolean hasExtendedRequestView = SpecimenManager.get().getExtendedSpecimenRequestView(context) != null;
     List<Vial> vials = bean.getVials();
     SpecimenRequestInput[] inputs = bean.getInputs();
 %>
