+++ study/src/org/labkey/study/view/specimen/specimen.jsp
@@ -15,16 +15,16 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.specimen.SpecimenManager"%>
 <%@ page import="org.labkey.api.specimen.Vial"%>
 <%@ page import="org.labkey.api.specimen.location.LocationManager"%>
-<%@ page import="org.labkey.api.specimen.model.SpecimenComment"%>
+<%@ page import="org.labkey.api.specimen.model.SpecimenComment" %>
 <%@ page import="org.labkey.api.specimen.security.permissions.SetSpecimenCommentsPermission" %>
 <%@ page import="org.labkey.api.study.Location" %>
 <%@ page import="org.labkey.api.study.StudyService" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.study.SpecimenManager" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
@@ -32,7 +32,7 @@
     SpecimenController.SpecimenEventBean bean = me.getModelBean();
     Vial vial = bean.getVial();
     Location originatingLocation = LocationManager.get().getOriginatingLocation(vial);
-    SpecimenComment comment = SpecimenManager.getInstance().getSpecimenCommentForVial(vial);
+    SpecimenComment comment = SpecimenManager.get().getSpecimenCommentForVial(vial);
     ActionURL commentsLink = new ActionURL(SpecimenController.UpdateCommentsAction.class, vial.getContainer());
     commentsLink.addParameter("rowId", vial.getRowId());
     commentsLink.addParameter("referrer", getActionURL().getLocalURIString());
